requirements.txt
asgiref==3.3.1
Django==3.1.6
django-crispy-forms==1.9.2
django-phone-field==1.8.1
django-sendgrid-v5==0.9.1
django-widget-tweaks==1.4.8
future==0.18.2
Pillow==8.1.0
python-http-client==3.3.2
pytz==2021.1
sendgrid==6.6.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
whitenoise==5.2.0
